feat: 레이아웃 조정 && 클라이언트, 서버 디렉터리 분리
