+++ client/package.json
@@ -12,6 +12,7 @@
   "dependencies": {
     "@emotion/babel-plugin": "^11.11.0",
     "@emotion/react": "^11.11.4",
+    "@emotion/styled": "^11.11.5",
     "@vitejs/plugin-react": "^4.3.1",
     "eslint-config-prettier": "^9.1.0",
     "eslint-plugin-prettier": "^5.1.3",
