+++ client/src/components/badge/Badge.tsx
