+++ client/src/components/badge/style.ts
