+++ client/src/components/badge/type.ts
