+++ client/src/components/log/style.ts
@@ -1,0 +1,21 @@
+import { css } from '@emotion/react';
+
+export const logStyle = css`
+  position: relative;
+  width: 20%;
+  min-width: 350px;
+  max-width: 400px;
+  min-height: 600px;
+  border: 1px solid gainsboro;
+  border-radius: 10px;
+  box-shadow: 0 0 20px 10px gainsboro;
+  left: -30px;
+
+  padding: 60px 40px;
+  padding-left: 70px;
+  .title {
+    font-size: 2.4rem;
+    font-weight: 900;
+    text-align: center;
+  }
+`;
