+++ client/src/style/style.tsx
@@ -1,0 +1,5 @@
+import styled from '@emotion/styled';
+
+export const Space = styled.div<{ size: number }>`
+  margin: ${p => p.size}px 0;
+`;
